+++ R/GraphLearner.R
@@ -212,8 +212,18 @@ as_learner.Graph = function(x, clone = FALSE) {
 }
 
 #' @export
+as_learners.Graph = function(x, clone = FALSE) {
+  list(as_learner(x))
+}
+
+#' @export
 as_learner.PipeOp = function(x, clone = FALSE) {
   as_learner(as_graph(x, clone = FALSE), clone = clone)
+}
+
+#' @export
+as_learners.PipeOp = function(x, clone = FALSE) {
+  list(as_learner(x))
 }